Bug Fixes, Information Validation, DM usage, Refactoring
Minor bug fix on wrong use of function.
All PromptInfo is validated!
Prompts can be used in DM channels!
Refactored the addition of text to the prompt when we are looking
for a specific amount. Was able to use broad wording to only need
the amount and nothing more. Also created a util function to create
the prompt using promptInfo.